+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 ### Added
 
 - Add `HandlerSet.Range[Aggregates|Processes|Integrations|Projections]()`
+- Add `RichHandlerSet.Range[Aggregates|Processes|Integrations|Projections]()`
 
 ## [0.2.2] - 2020-01-29